// src/app/interfaces/empleado.ts
// Generated by https://quicktype.io

export interface Empleado {
    EmployeeNumber: number;
    Password:       string;
    GivenName:      string;
    LastName:       string;
    Function:       number;
    Access:         string;
    Status:         boolean;
    RecordDate:     string;
    UpdateDate:     string;
}